Change the structure of the files
Previously, the directory structure would have been one directory per
sample, and then rely on the names to sort it out.  However, it seems
much cleaner to have the directory structure based on one directory per
fruiting body, with sub-directories for stalks and spores.

Also added in dummy entries in the config file for the sequencing data,
which should be coming off the sequencer really soon.