Fix the bug with hyper-inflated p-values
The problem is basically that there are lots of SNPs where I have no
alternative allele counts.  Those are throwing off the "expected"
p-value distributions.  Plus, there also turned out to be a bug in how I
was detecting which SNPs were stalk/spore biased (I did odds_ratio < 0,
instead of odds_ratio <1, thinking that it would be log odds ratio),
which couldn't have helped.

So now I only count a SNP in a given sample if it has any reads in the
alternative allele (for completeness, I also check if it has reads in
the reference allele).  This brings down the number of SNPs I'm looking
at to 14,099 instead of 183,712. In turn, that gives me a totally
reasonable looking QQ plot.