Switch back to BAM files
Picard seems unhappy taking CRAM files as input, so screw 'em.  I'll
just stick with BAM for the moment.

In the long run, probably would make sense to think about marking bam
files as temp and converting to cram for storage. But space isn't a huge
premium at the moment (though perhaps more so on sherlock)...