More poking around with MCMC
It seems that Pymc3 doesn't do well with binary variables. Try out stan,
but having a hard time getting that to work either.

One possibility Mark pointed out is instead of having a binary variable,
just have a Beta distributed one with small alpha and beta, and that way
it's almost binary.