More flexible way to specify files
Previously, it was relatively hard coded where to look for files.
Especially as we get data from multiple runs and multiple vendors, it
seems likely that we'll want to be fairly obvious about how we specify
what the files look like.

New approach is to have a list of formats, and it will loop through and
take files that are unique.  One should take some care to have it be
specific enough that it doesn't specify multiple files incorrectly.