Include both active and inactive samples
Active samples are those that will go into the final calculation of SNP
p-values, whereas inactive ones are those that should go through most of
the pipeline, but not feed into the last step. These latter samples are,
for instance, the pilot samples I generated in August/September that are
very contaminated, or a sample from the November run where I missed
grabbing a stalk.