Plot expression of genes near significant SNPs
Presumably these SNPs act through gene expression.  I don't know that we
know enough about the regulatory logic to directly guess at what genes
are being regulated, but I think it's a fair* assumption that whatever
genes the SNPs control will be close to the SNPs.

To do this, I also need to parse out and intersect bed files to find
genes that are near to the SNPs of interest.

* by fair I mean both in the sense of reasonable, and also "neither
excellent nor poor; moderately or tolerably good"