Add in code that makes various coverage graphs
Following up the observation that most of my variants were in genes, it
turned out that most of my reads were in genes, and the most likely
reason for this is that:
a) The dicty genome is AT rich
b) Very AT rich regions aren't getting amplified
c) Genes are slightly less AT rich, so that's where the reads are.

So really I need to update my library prep procedures (and there are
ways to improve AT rich sequencing), but this is a bolus of code that
helped me come to that conclusion.