Add option to plot LD for all pairs of SNPs at a given distance
In what I think was a moment of mild inspiration, I came up with a
reasonably elegant way to keep track of all SNPs that are a given
distance apart.

I do this by keeping a double-ended queue (dequeue), and push on the
right when I add a new SNP, and pop off the left when the distance gets
too big.

This has the advantage of increasing the total amount of data, but I
think it's still pretty reasonable to justify what's going on here.