Create bedfiles for SNPs that affect GC content
We could imagine (though not necessarily expect, given the style of
experiment) that there would be a difference in how well we are able to
recover SNPs that make things more AT/GC balanced. So this is a quick
rule to break the autosomal SNPs into three classes (more, less, or same
GC content).