Fix bug with automatically creating GFF file
Unzip returns a non-zero error code if we are using the -j flag, so the
script cannot run. To get around this, I put `set +e` in front of it to
suppress the strict mode, then reenable it when the unzipping is done.

Also, split downloading the file into its own rule to avoid unnecessary
repeats.